Corrected js file
