+++ src/js/variables/constants.js
@@ -1,4 +1,6 @@
-export const URL = 'https://api.themoviedb.org/3/';
+export const API_URL = 'https://api.themoviedb.org/3/';
 export const IMG_URL = 'https://image.tmdb.org/t/p/w500';
 export const API_KEY = 'e56bac74c49b45071b964a5f7a0906aa';
 export const LANGUAGE = 'en-US';
+
+export default { API_URL, IMG_URL, API_KEY, LANGUAGE }
